# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Before Bringing It Up API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/scenarios/{scenario_id}": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "Get Scenario",
        "operationId": "get_scenario_api_scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/validate": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Validate Scenarios",
        "operationId": "validate_scenarios_api_scenarios_validate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Scenarios Api Scenarios Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunRecord"
                  },
                  "title": "Response List Runs Api Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Compare",
        "operationId": "compare_api_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunRecord"
                  },
                  "type": "array",
                  "title": "Response Compare Api Compare Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/study/sessions": {
      "post": {
        "tags": [
          "study"
        ],
        "summary": "Create Session",
        "operationId": "create_session_api_study_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/study/sessions/{session_id}/next": {
      "get": {
        "tags": [
          "study"
        ],
        "summary": "Next Assignment",
        "operationId": "next_assignment_api_study_sessions__session_id__next_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Next Assignment Api Study Sessions  Session Id  Next Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/study/sessions/{session_id}/responses": {
      "post": {
        "tags": [
          "study"
        ],
        "summary": "Add Response",
        "operationId": "add_response_api_study_sessions__session_id__responses_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyRating"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Add Response Api Study Sessions  Session Id  Responses Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/study/sessions/{session_id}/withdraw": {
      "post": {
        "tags": [
          "study"
        ],
        "summary": "Withdraw",
        "operationId": "withdraw_api_study_sessions__session_id__withdraw_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Withdraw Api Study Sessions  Session Id  Withdraw Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/export/runs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Export Runs",
        "operationId": "export_runs_api_admin_export_runs_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompareRequest": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "provider": {
            "type": "string",
            "const": "mock",
            "title": "Provider",
            "default": "mock"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 454491
          }
        },
        "type": "object",
        "required": [
          "scenario_id"
        ],
        "title": "CompareRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MethodName": {
        "type": "string",
        "enum": [
          "no_memory",
          "similarity_top_k",
          "one_pass_selective",
          "relevance_two_pass",
          "reconsider_lite",
          "no_physical_separation"
        ],
        "title": "MethodName"
      },
      "ParticipantProfile": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "description"
        ],
        "title": "ParticipantProfile"
      },
...
      "RunRecord": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id",
            "default": "adhoc"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "scenario_version": {
            "type": "integer",
            "title": "Scenario Version"
          },
          "method": {
            "$ref": "#/components/schemas/MethodName"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "prompt_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Prompt Versions"
          },
          "prompt_hashes": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Prompt Hashes"
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "random_seed": {
            "type": "integer",
            "title": "Random Seed"
          },
          "candidate_order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Candidate Order"
          },
          "hard_gates": {
            "$ref": "#/components/schemas/GateBundle"
          },
          "deliberation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeliberationBundle"
              },
              {
                "type": "null"
              }
            ]
          },
          "controller_decisions": {
            "items": {
              "$ref": "#/components/schemas/ControllerDecision"
            },
            "type": "array",
            "title": "Controller Decisions"
          },
          "admitted_views": {
            "items": {
              "$ref": "#/components/schemas/AdmittedMemoryView"
            },
            "type": "array",
            "title": "Admitted Views"
          },
          "visible_reply": {
            "type": "string",
            "title": "Visible Reply"
          },
          "actions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PublicAction"
            },
            "type": "object",
            "title": "Actions"
          },
          "validator_issues": {
            "items": {
              "$ref": "#/components/schemas/ValidatorIssue"
            },
            "type": "array",
            "title": "Validator Issues"
          },
          "repair_count": {
            "type": "integer",
            "title": "Repair Count",
            "default": 0
          },
          "fallback_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Type"
          },
          "latency": {
            "$ref": "#/components/schemas/StageLatency"
          },
          "token_usage": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Token Usage"
          },
          "configured_cost_estimate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configured Cost Estimate"
          },
          "schema_valid": {
            "type": "boolean",
            "title": "Schema Valid",
            "default": true
          },
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "output_hash": {
            "type": "string",
            "title": "Output Hash"
          },
          "generator_request_json": {
            "type": "string",
            "title": "Generator Request Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "software_commit_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Software Commit Hash"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "scenario_id",
          "scenario_version",
          "method",
          "provider",
          "model_id",
          "config_hash",
          "random_seed",
          "candidate_order",
          "hard_gates",
          "visible_reply",
          "input_hash",
          "output_hash",
          "generator_request_json"
        ],
        "title": "RunRecord"
      },
      "RunRequest": {
        "properties": {
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "scenario": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Scenario"
              },
              {
                "type": "null"
              }
            ]
          },
          "method": {
            "$ref": "#/components/schemas/MethodName",
            "default": "reconsider_lite"
          },
          "provider": {
            "type": "string",
            "const": "mock",
            "title": "Provider",
            "default": "mock"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 454491
          }
        },
        "type": "object",
        "title": "RunRequest"
      },
      "Scenario": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "family_id": {
            "type": "string",
            "title": "Family Id"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "set": {
            "type": "string",
            "title": "Set",
            "default": "core"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "participant_profile": {
            "$ref": "#/components/schemas/ParticipantProfile"
          },
          "conversation": {
            "$ref": "#/components/schemas/ConversationInput"
          },
          "gold": {
            "$ref": "#/components/schemas/GoldAnnotations"
          },
          "notes": {
            "$ref": "#/components/schemas/ScenarioNotes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_id",
          "family_id",
          "version",
          "title",
          "participant_profile",
          "conversation",
          "gold"
        ],
        "title": "Scenario"
      },
...
      "StudyRating": {
        "properties": {
          "assignment_id": {
            "type": "string",
            "title": "Assignment Id"
          },
          "relational_appropriateness": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Relational Appropriateness"
          },
          "helpfulness": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Helpfulness"
          },
          "naturalness": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Naturalness"
          },
          "continuity": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Continuity"
          },
          "feeling_understood": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Feeling Understood"
          },
          "intrusion": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Intrusion"
          },
          "creepiness": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Creepiness"
          },
          "privacy_concern": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Privacy Concern"
          },
          "trust": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Trust"
          },
          "user_agency": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "User Agency"
          },
          "rationale": {
            "type": "string",
            "maxLength": 1000,
            "title": "Rationale",
            "default": ""
          },
          "skipped": {
            "type": "boolean",
            "title": "Skipped",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "assignment_id",
          "relational_appropriateness",
          "helpfulness",
          "naturalness",
          "continuity",
          "feeling_understood",
          "intrusion",
          "creepiness",
          "privacy_concern",
          "trust",
          "user_agency"
        ],
        "title": "StudyRating"
      },
      "StudySession": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "participant_code": {
            "type": "string",
            "title": "Participant Code"
          },
          "protocol_version": {
            "type": "string",
            "title": "Protocol Version"
          },
          "consented_at": {
            "type": "string",
            "format": "date-time",
            "title": "Consented At"
          },
          "assignment_seed": {
            "type": "integer",
            "title": "Assignment Seed"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "participant_code",
          "protocol_version",
          "consented_at",
          "assignment_seed"
        ],
        "title": "StudySession"
      },
      "StudySessionCreate": {
        "properties": {
          "adult_eligible": {
            "type": "boolean",
            "title": "Adult Eligible"
          },
          "consented": {
            "type": "boolean",
            "title": "Consented"
          },
          "protocol_version": {
            "type": "string",
            "title": "Protocol Version"
          }
        },
        "type": "object",
        "required": [
          "adult_eligible",
          "consented",
          "protocol_version"
        ],
        "title": "StudySessionCreate"
      },
...
    }
  }
}